# openapi.yaml
openapi: 3.0.3
info:
  title: Payment Management API
  description: |
    API for managing payments, vendors, items, and invoices.
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. 
    Obtain a token by calling the `/auth/login` endpoint.
    
    ## Filtering
    The Vendors and Items endpoints support advanced filtering with the following operators:
    - `=` - Exact match
    - `contains` - Fuzzy search (case-insensitive)
    
    ## Pagination
    All list endpoints support pagination with the following parameters:
    - `page` - Page number (default: 1)
    - `limit` - Items per page (default: 10, max: 100)
    
    Responses include pagination metadata with total count, page info, and navigation flags.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:3000/api
    description: Local development server

tags:
  - name: Authentication
    description: User authentication and registration
  - name: Vendors
    description: Vendor management operations
  - name: Items
    description: Item management operations
  - name: Invoices
    description: Invoice management operations

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: admin@example.com
        name:
          type: string
          example: Admin User
      required:
        - id
        - email

    Vendor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: TechCorp
        contact:
          type: string
          nullable: true
          example: contact@techcorp.com
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
      required:
        - id
        - name

    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Laptop
        price:
          type: number
          format: float
          example: 1200.00
        vendorId:
          type: integer
          example: 1
        vendor:
          $ref: '#/components/schemas/VendorBasic'
      required:
        - id
        - name
        - price
        - vendorId

    VendorBasic:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: TechCorp
        contact:
          type: string
          nullable: true
          example: contact@techcorp.com

    Invoice:
      type: object
      properties:
        id:
          type: integer
          example: 1
        date:
          type: string
          format: date-time
          example: 2025-11-21T10:00:00Z
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
          example: PENDING
        totalAmount:
          type: number
          format: float
          example: 2527.50
        userId:
          type: integer
          nullable: true
          example: 1
        user:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItem'

    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        invoiceId:
          type: integer
          example: 1
        itemId:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 1200.00
        item:
          $ref: '#/components/schemas/Item'

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials

    PaginationMetadata:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 50
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 5
        hasNext:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPrevious:
          type: boolean
          description: Whether there is a previous page
          example: false

    PaginatedVendorsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vendor'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'

    PaginatedItemsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'

    PaginatedInvoicesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Test User
                email:
                  type: string
                  format: email
                  example: test@example.com
                password:
                  type: string
                  format: password
                  example: password123
              required:
                - email
                - password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
                password:
                  type: string
                  format: password
                  example: password123
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset instructions to user's email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset email sent

  /vendors:
    get:
      tags:
        - Vendors
      summary: List all vendors
      description: |
        Retrieve a list of vendors with optional filtering and pagination.
        
        **Filtering Examples:**
        - Get vendor by ID: `?id=1&idOperator==`
        - Fuzzy search by name: `?name=tech&nameOperator=contains`
        - Exact name match: `?name=TechCorp&nameOperator==`
        
        **Pagination Examples:**
        - Get first page: `?page=1&limit=10`
        - Get second page with 20 items: `?page=2&limit=20`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: Filter by vendor ID
          schema:
            type: integer
            example: 1
        - name: idOperator
          in: query
          description: Operator for ID filtering
          schema:
            type: string
            enum: ['=', 'contains']
            default: '='
        - name: name
          in: query
          description: Filter by vendor name
          schema:
            type: string
            example: tech
        - name: nameOperator
          in: query
          description: Operator for name filtering
          schema:
            type: string
            enum: ['=', 'contains']
            default: contains
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Paginated list of vendors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVendorsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Vendors
      summary: Create a new vendor
      description: Add a new vendor to the system
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: New Vendor
                contact:
                  type: string
                  example: contact@newvendor.com
              required:
                - name
      responses:
        '200':
          description: Vendor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '401':
          description: Unauthorized
        '500':
          description: Failed to create vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /items:
    get:
      tags:
        - Items
      summary: List all items
      description: |
        Retrieve a list of items with optional filtering by vendor and pagination.
        
        **Filtering Examples:**
        - Get items by vendor ID: `?vendorId=1&vendorIdOperator==`
        - Fuzzy search by vendor name: `?vendorName=tech&vendorNameOperator=contains`
        - Exact vendor name match: `?vendorName=TechCorp&vendorNameOperator==`
        
        **Pagination Examples:**
        - Get first page: `?page=1&limit=10`
        - Get second page with 20 items: `?page=2&limit=20`
      security:
        - bearerAuth: []
      parameters:
        - name: vendorId
          in: query
          description: Filter by vendor ID
          schema:
            type: integer
            example: 1
        - name: vendorIdOperator
          in: query
          description: Operator for vendor ID filtering
          schema:
            type: string
            enum: ['=']
            default: '='
        - name: vendorName
          in: query
          description: Filter by vendor name
          schema:
            type: string
            example: tech
        - name: vendorNameOperator
          in: query
          description: Operator for vendor name filtering
          schema:
            type: string
            enum: ['=', 'contains']
            default: contains
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Paginated list of items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedItemsResponse'
        '401':
          description: Unauthorized

    post:
      tags:
        - Items
      summary: Create a new item
      description: Add a new item to the system
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Keyboard
                price:
                  type: number
                  format: float
                  example: 75.00
                vendorId:
                  type: integer
                  example: 1
              required:
                - name
                - price
                - vendorId
      responses:
        '200':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
        '500':
          description: Failed to create item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /invoices:
    get:
      tags:
        - Invoices
      summary: List all invoices
      description: |
        Retrieve a paginated list of all invoices with items and user information.
        Invoices are ordered by date (most recent first).
        
        **Pagination Examples:**
        - Get first page: `?page=1&limit=10`
        - Get second page with 20 items: `?page=2&limit=20`
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Paginated list of invoices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvoicesResponse'
        '401':
          description: Unauthorized

    post:
      tags:
        - Invoices
      summary: Create a new invoice
      description: Create a new invoice with line items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      itemId:
                        type: integer
                        example: 1
                      quantity:
                        type: integer
                        example: 2
                      price:
                        type: number
                        format: float
                        example: 1200.00
                    required:
                      - itemId
                      - quantity
                      - price
                userId:
                  type: integer
                  nullable: true
                  description: Optional user ID (defaults to authenticated user)
                  example: 1
              required:
                - items
      responses:
        '200':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '500':
          description: Failed to create invoice

  /invoices/{id}/approve:
    put:
      tags:
        - Invoices
      summary: Approve an invoice
      description: Change invoice status to APPROVED
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Invoice ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Invoice approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          description: Unauthorized
        '500':
          description: Failed to approve invoice

  /invoices/{id}/reject:
    put:
      tags:
        - Invoices
      summary: Reject an invoice
      description: Change invoice status to REJECTED
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Invoice ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Invoice rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          description: Unauthorized
        '500':
          description: Failed to reject invoice